Failed attept at creating CustomSliderInput story
Ultimently I gave this up because for CustomSilderInput to be fully
tested, it needs the openDialog to be tested that occurs when the value
is clicked. Clicking this dispatches an action to the redux store. This
store then updates the App CustomDialog state to have a prompt where the
input will get back to the slider.

These files try to seperate the App into a Wrapper file, such that it
could be used as a Storybook decoration. However, this does not work
because the App relies on non-GUI functions like "window.require"? And
accessing parts of the OS which a story should definently not do.

In conclusion, redux connected components can probably not be tested
easily and stories should only be made for static components.